requirements.txt
# Core Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Deep Learning and Computer Vision
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.1.78
Pillow==10.0.1

# Audio Processing
librosa==0.10.1
soundfile==0.12.1

# Machine Learning
scikit-learn==1.3.0
joblib==1.3.2

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Utilities
tqdm==4.66.1
matplotlib==3.7.2
seaborn==0.12.2

# Jupyter Notebook (for training)
jupyter==1.0.0
ipykernel==6.25.2

# Additional Audio Dependencies
kagglehub==0.2.5
ffmpeg-python==0.2.0

# Optional: CUDA support (uncomment if using GPU)
# torch==2.0.1+cu118
# torchvision==0.15.2+cu118
# --extra-index-url https://download.pytorch.org/whl/cu118